:sparkles: PagingDataSource의 Firebase 의존성 제거
