+++ domain/src/main/java/com/whyranoid/domain/repository/PostRepository.kt
@@ -2,13 +2,14 @@ package com.whyranoid.domain.repository
 
 import androidx.paging.PagingData
 import com.whyranoid.domain.model.Post
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 
 interface PostRepository {
 
     // TODO : 페이징 처리
     // 글(홍보 / 인증) 페이징으로 가져오기 - 리모트
-    fun getPagingPosts(): Flow<PagingData<Post>>
+    fun getPagingPosts(coroutineScope: CoroutineScope): Flow<PagingData<Post>>
 
     fun getMyPagingPosts(uid: String): Flow<PagingData<Post>>
